pyproject.toml
[project]
name = "openai-agents-fastapi-playbook"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiosqlite>=0.20.0",
    "autoevals>=0.0.130",
    "braintrust>=0.3.1",
    "dspy>=3.0.3",
    "fastapi[standard]>=0.119.0",
    "ipykernel>=7.0.1",
    "langchain-text-splitters>=1.1.0",
    "matplotlib>=3.10.8",
    "openai>=2.6.1",
    "openai-agents>=0.4.2",
    "pandas>=2.3.3",
    "pycryptodome>=3.23.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.2",
    "tavily-python>=0.7.12",
    "tiktoken>=0.12.0",
]
